Move Hashable trait to hashing module
