+++ src/lib.rs
@@ -25,24 +25,11 @@
 //! roll back to any previous checkpoint.
 
 pub mod bridgetree;
+pub mod hashing;
 pub mod position;
 
-use position::{Level, Position};
+use position::Position;
 use std::collections::BTreeSet;
-
-/// A trait describing the operations that make a value  suitable for inclusion in
-/// an incremental merkle tree.
-pub trait Hashable: Sized {
-    fn empty_leaf() -> Self;
-
-    fn combine(level: Level, a: &Self, b: &Self) -> Self;
-
-    fn empty_root(level: Level) -> Self {
-        Level::from(0)
-            .iter_to(level)
-            .fold(Self::empty_leaf(), |v, lvl| Self::combine(lvl, &v, &v))
-    }
-}
 
 /// A possibly-empty incremental Merkle frontier.
 pub trait Frontier<H> {
